+++ nix/launcher-config.nix
@@ -26,11 +26,6 @@ let
       cluster = "mainnet";
       networkName = "mainnet";
     };
-    alonzo_purple = {
-      cardanoEnv = cardanoLib.environments.alonzo-purple;
-      cluster = "alonzo-purple";
-      networkName = "alonzo-purple";
-    };
   };
   dirSep = if os == "windows" then "\\" else "/";
   configDir = configFilesSource: {