+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Changelog
 
 ### Chores
 
-- Updated `cardano-wallet` to version `2021-07-30` ([PR 2635](https://github.com/input-output-hk/daedalus/pull/2635))
+- Updated `cardano-wallet` to version `2021-07-30` ([PR 2635](https://github.com/input-output-hk/daedalus/pull/2635), [PR 2637](https://github.com/input-output-hk/daedalus/pull/2637))
 - Removed "Decentralization countdown" tada icon and info tab ([PR 2625](https://github.com/input-output-hk/daedalus/pull/2625))
 
 ## 4.2.0